# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Event Booking Platform API",
    "description": "Backend API for an event booking and ticketing platform. All endpoints return `{ success: true, data: ... }` on success and `{ success: false, error: '...' }` on failure.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/api/v1", "description": "API v1 (recommended)" },
    { "url": "/api", "description": "Legacy /api prefix (deprecated)" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/admin/uploads/event": {
      "post": {
        "tags": ["Uploads"],
        "summary": "Upload event image",
        "security": [{ "BearerAuth": [] }],
        "responses": { "201": { "description": "File uploaded" } }
      }
    },
    "/admin/uploads/banner": {
      "post": {
        "tags": ["Uploads"],
        "summary": "Upload banner image",
        "security": [{ "BearerAuth": [] }],
        "responses": { "201": { "description": "File uploaded" } }
      }
    }
  }
}